backend/api-server/services/requirements.txt
pywhatkit
selenium
webdriver-manager
python-dotenv
